refactor: Add modularisation to notes and folders related screens
