+++ app/src/main/java/com/github/onlynotesswent/ui/overview/FolderContent.kt
@@ -2,6 +2,7 @@ package com.github.onlynotesswent.ui.overview
 
 import android.util.Log
 import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
@@ -14,10 +15,13 @@ import androidx.compose.material.icons.filled.Add
 import androidx.compose.material.icons.filled.Clear
 import androidx.compose.material.icons.filled.Edit
 import androidx.compose.material.icons.filled.MoreVert
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Button
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
@@ -45,7 +49,6 @@ import com.github.onlynotesswent.ui.CreateFolderDialog
 import com.github.onlynotesswent.ui.CustomDropDownMenu
 import com.github.onlynotesswent.ui.CustomDropDownMenuItem
 import com.github.onlynotesswent.ui.CustomLazyGrid
-import com.github.onlynotesswent.ui.RenameFolderDialog
 import com.github.onlynotesswent.ui.navigation.NavigationActions
 import com.github.onlynotesswent.ui.navigation.Screen
 import com.github.onlynotesswent.ui.navigation.TopLevelDestinations
@@ -435,4 +438,43 @@ fun FolderContentScreenGrid(
             )
         }
     )
+}
+
+/**
+ * Dialog that allows the user to rename a folder.
+ *
+ * @param currentName the current name of the folder
+ * @param onDismiss callback to be invoked when the dialog is dismissed
+ * @param onConfirm callback to be invoked when the user confirms the new name
+ */
+@Composable
+fun RenameFolderDialog(currentName: String, onDismiss: () -> Unit, onConfirm: (String) -> Unit) {
+
+    var newName by remember { mutableStateOf(currentName) }
+
+    AlertDialog(
+        modifier = Modifier.testTag("renameFolderDialog"),
+        onDismissRequest = onDismiss,
+        title = {
+            Box(modifier = Modifier.fillMaxWidth(), contentAlignment = Alignment.Center) {
+                Text("Rename folder")
+            }
+        },
+        text = {
+            OutlinedTextField(
+                value = newName, onValueChange = { newName = it }, label = { Text("New Folder Name") })
+        },
+        confirmButton = {
+            Button(
+                enabled = newName.isNotEmpty(),
+                onClick = { onConfirm(newName) },
+                modifier = Modifier.testTag("confirmRenameButton")) {
+                Text("Confirm")
+            }
+        },
+        dismissButton = {
+            Button(onClick = onDismiss, modifier = Modifier.testTag("dismissRenameButton")) {
+                Text("Cancel")
+            }
+        })
 }
